examples/vae_mnist/pyproject.toml
[project]
name = "vae-mnist-distributed"
version = "0.1.0"
description = "Distributed VAE training on MNIST across Apple Silicon Macs using MCCL"
requires-python = ">=3.11"
dependencies = [
    "torch>=2.5.0",
    "torchvision>=0.20.0",
    "distribute-metal",
    "pyyaml>=6.0",
]

[tool.uv.sources]
distribute-metal = { path = "../../packages/distribute_metal" }
